fix: resolve exact formatting mismatches for e2e test script

+++ src/main/kotlin/com/example/financemanager/dto/response/GoalResponses.kt
@@ -10,7 +10,7 @@ data class GoalDto(
     val targetDate: LocalDate,
     val startDate: LocalDate,
     val currentProgress: BigDecimal,
-    val progressPercentage: BigDecimal,
+    val progressPercentage: Double,
     val remainingAmount: BigDecimal
 )
